Revert "Merge branch 'ppy:master' into no-combo-scaling"
This reverts commit 68a6682c2bc3574752998a9d8329c9ac4b533a13, reversing
changes made to e2a5d1904b000965f009ee746a3175396011e3fc.

+++ osu.Game/Online/Rooms/IndexedMultiplayerScores.cs
@@ -17,7 +17,7 @@ namespace osu.Game.Online.Rooms
         /// The total scores in the playlist item.
         /// </summary>
         [JsonProperty("total")]
-        public long? TotalScores { get; set; }
+        public int? TotalScores { get; set; }
 
         /// <summary>
         /// The user's score, if any.
